Feat: Completed destination Crud, Partially finished role based login system

+++ TravelTalesBackend/app/routes/destination_route.py
@@ -1,4 +1,7 @@
 from typing import List, Optional
+from app.auth.auth import get_current_user
+from app.model.models import User
+from fastapi.exceptions import RequestValidationError
 from fastapi import APIRouter, Depends, HTTPException, Query, status
 from sqlalchemy.orm import Session
 from app.utils.db_utils import get_db
@@ -6,7 +9,8 @@ from app.services.destination_service import (
     create_destination as db_create_destination, 
     get_destination_by_id, 
     get_all_destinations, 
-    delete_destination)
+    delete_destination,
+    update_destination_service)
 
 
 from app.schemas.schemas import(
@@ -25,11 +29,22 @@ router = APIRouter(
 @router.post("/", response_model= DestinationResponse, status_code= status.HTTP_201_CREATED)
 def create_destination(
     destination : DestinationCreate,
-    db: Session = Depends(get_db)
+    current_user: User = Depends(get_current_user),
+    db: Session = Depends(get_db),
 ):
+    if current_user.roles != "admin":
+        raise HTTPException(
+            status_code= status.HTTP_403_FORBIDDEN,
+            detail= "Only admin can create new destination"
+        )
     try: 
         new_destination = db_create_destination(db, destination)
         return new_destination
+    except RequestValidationError as ve:
+        raise HTTPException(
+            status_code= status.HTTP_400_BAD_REQUEST,
+            detail= ve.errors()
+        )
     except Exception as e:
         raise HTTPException(
             status_code= status.HTTP_500_INTERNAL_SERVER_ERROR,
@@ -82,8 +97,15 @@ def get_destination_id(
 @router.delete("/{destination_id}", status_code= status.HTTP_204_NO_CONTENT)
 def delete_destination_route(
     destination_id: int,
+    current_user: User = Depends(get_current_user),
     db: Session = Depends(get_db)
 ):
+    
+    if current_user.roles != "admin":
+        raise HTTPException(
+            status_code= status.HTTP_403_FORBIDDEN,
+            detail= "Only admin can delete destination"
+        )
     try:
         destination = delete_destination(db, destination_id)
 
@@ -100,3 +122,25 @@ def delete_destination_route(
             status_code=status.HTTP_500_INTERNAL_SERVER_ERROR,
             detail="An unexpected error occurred"
         ) 
+
+@router.patch("/}", response_model = DestinationResponse)
+def update_destination(
+    destination_update: DestinationUpdate,
+    destination_id : int,
+    current_user : User = Depends(get_current_user),
+    db: Session = Depends(get_db)
+):
+    if current_user.roles != "admin":
+        raise HTTPException(
+            status_code = status.HTTP_403_FORBIDDEN,
+            detail= "Only admin is allowed to update the information"
+        )
+    update_destination = update_destination_service(
+        db, destination_id, destination_update
+    )
+    if not update_destination:
+        raise HTTPException(
+            status_code= status.HTTP_404_NOT_FOUND,
+            detail= f"Destination with id {destination_id} not found"
+        )
+    return update_destination
